Move liveness transfer function to new header
